+++ src/test/kotlin/util/NumericValidatorTest.kt
@@ -1,9 +1,8 @@
-package utiltest
+package util
 
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import util.NumericValidator
 
 class NumericValidatorTest {
